requirements.txt
python-dotenv==1.1.1
requests==2.32.5
pydantic==2.11.7
click==8.2.1
tenacity==9.1.2
pandas==2.3.2
pytest==8.4.1
pytest-cov==6.2.1
pyflakes==3.4.0
black==25.1.0
tqdm==4.66.5
